+++ src/components/ExportPanel.svelte
@@ -1,9 +1,9 @@
 <script lang="ts">
-    import Panel from './basic/Panel.svelte'
-    import Range from './basic/Range.svelte'
-    import Checkbox from './basic/Checkbox.svelte'
-    import Button from './basic/Button.svelte'
-    import TextField from './basic/TextField.svelte'
+    import Panel from './Panel.svelte'
+    import Range from './Range.svelte'
+    import Checkbox from './Checkbox.svelte'
+    import Button from './Button.svelte'
+    import TextField from './TextField.svelte'
     import { fundamental, sampleDuration, sampleName, partials, synthPartials, dissonanceCurve, dissCurveLimits, notes } from '../state/stores.js'
     import { BlobReader, BlobWriter, TextReader, ZipWriter } from '@zip.js/zip.js'
     import { AdditiveSynth } from '../xentonality/synth'
