+++ src/components/Footer.svelte
@@ -1,11 +1,11 @@
 <script lang="ts">
-    import KoFiIcon from './icons/KoFiIcon.svelte'
-    import YouTubeIcon from './icons/YouTubeIcon.svelte'
-    import TwitterIcon from './icons/TwitterIcon.svelte'
-    import MailIcon from './icons/MailIcon.svelte'
-    import MenuItem from './basic/MenuItem.svelte'
-    import BandcampIcon from './icons/BandcampIcon.svelte'
-    import GithubIcon from './icons/GithubIcon.svelte'
+    import KoFiIcon from '../icons/KoFiIcon.svelte'
+    import YouTubeIcon from '../icons/YouTubeIcon.svelte'
+    import TwitterIcon from '../icons/TwitterIcon.svelte'
+    import MailIcon from '../icons/MailIcon.svelte'
+    import MenuItem from './MenuItem.svelte'
+    import BandcampIcon from '../icons/BandcampIcon.svelte'
+    import GithubIcon from '../icons/GithubIcon.svelte'
 </script>
 
 <footer class="w-full absolute left-0 bottom-0 flex justify-center bg-white-5 py-4 border-t border-white-65">
