+++ src/components/Panel.svelte
