+++ src/components/Range.svelte
