+++ src/icons/BandcampIcon.svelte
