+++ src/icons/CrossIcon.svelte
