+++ src/icons/DownloadIcon.svelte
