+++ src/icons/MailIcon.svelte
