+++ src/icons/SettingsIcon.svelte
