+++ src/icons/YouTubeIcon.svelte
